Create Lab2 Exercise 2
Cod was created to light up 3 different LED lights with just one switch controlling their behaviour depending if it receives voltage from the Arduino or not (on/off). Light begin to consecutively light up one after another (depending on which one was coded first from up to down), the blinking behaviours will be based on the delay that has been set to them. So it is regular one second speed when switch is inactive and is the personal blinking speed per LED while the switch is active.