Create Lab 2 Exercise 3
This code is just made to show avoid confusion for myself if the DIY switch that I made for this Lab worked. I made the LED lights turn on with no blink, though one of them is directly affected by the DIY switch that I made, it shuts down once the Switch goes off.